+++ packages/plugin-packages/mashroom-portal/src/frontend/portal-client/js/MashroomPortalRemoteLoggerImpl.ts
@@ -2,12 +2,15 @@
 import {WINDOW_VAR_PORTAL_API_PATH} from '../../../backend/constants';
 
 import type {LogLevel} from '@mashroom/mashroom/type-definitions';
-import type {MashroomPortalRemoteLogger, MashroomRestService} from '../../../../type-definitions';
+import type {
+    MashroomRestService,
+    MasterMashroomPortalRemoteLogger
+} from '../../../../type-definitions';
 import type {ClientLogMessage} from '../../../../type-definitions/internal';
 
 const SEND_INTERVAL = 1000;
 
-export default class MashroomPortalRemoteLoggerImpl implements MashroomPortalRemoteLogger {
+export default class MashroomPortalRemoteLoggerImpl implements MasterMashroomPortalRemoteLogger {
 
     private _restService: MashroomRestService;
     private _unsentLogMessages: Array<ClientLogMessage>;
@@ -20,11 +23,11 @@ export default class MashroomPortalRemoteLoggerImpl implements MashroomPortalRem
         this._timeout = null;
     }
 
-    error(msg: string, error?: Error, portalAppName?: string | undefined | null) {
+    error(msg: string, error?: Error, portalAppName?: string | null | undefined) {
         this._log('error', msg, error, portalAppName);
     }
 
-    warn(msg: string, error?: Error, portalAppName?: string | undefined | null) {
+    warn(msg: string, error?: Error, portalAppName?: string | null | undefined) {
         this._log('warn', msg, error, portalAppName);
     }
 
@@ -34,19 +37,16 @@ export default class MashroomPortalRemoteLoggerImpl implements MashroomPortalRem
 
     getAppInstance(portalAppName: string) {
         // eslint-disable-next-line @typescript-eslint/no-this-alias
-        const self = this;
+        const master = this;
         return {
             error(msg: string, error?: Error) {
-                self.error(msg, error, portalAppName);
+                master.error(msg, error, portalAppName);
             },
             warn(msg: string, error?: Error) {
-                self.error(msg, error, portalAppName);
+                master.warn(msg, error, portalAppName);
             },
             info(msg: string) {
-                self.info(msg, portalAppName);
-            },
-            getAppInstance() {
-                throw Error('Not implemented');
+                master.info(msg, portalAppName);
             }
         }
     }
@@ -86,8 +86,12 @@ export default class MashroomPortalRemoteLoggerImpl implements MashroomPortalRem
         }
     }
 
-    private _serializeError(error: Error) {
-        const msg = JSON.stringify({...error}, null, 2);
+    private _serializeError(error: any) {
+        const errorObj: any = {};
+        Object.getOwnPropertyNames(error).forEach((key) => {
+            errorObj[key] = error[key];
+        }, this);
+        const msg = JSON.stringify(errorObj, null, 2);
         return msg.replace(/\\n/g, '\n');
     }
 }