+++ CHANGELOG.md
@@ -7,7 +7,9 @@
  * Sandbox App: The sandbox now uses the "real" message bus and therefore supports Apps that use the messageBus
    interceptor, and can also simulate communication with other Apps on the same page.
    Also, the Sandbox can now load itself properly ;-)
- * Portal: Added an info() method to *MashroomPortalRemoteLogger*
+ * Portal: MashroomPortalRemoteLogger improved:
+     * Added an info() method
+     * Fixed client side error serialisation
  * Virtual host path mapper: Make sure the reverse mapped location header is never empty (fixes #79)
 
 ## 1.7.8 (May 31, 2021)